Fixed a bug when adding a link to an empty collection
If having a page with an empty link collection, the model for the collection will simply be an empty array. However when saving the page, Umbraco will change the model to an empty string, which doesn't have the `push` method and similar features of a JavaScript array, leading to a JavaScript error and the link not being properly added to the link collection.

Reloading the page will fix this as the controller of the Angular directive will be called again.

With this commit, we now set the model to an empty array (if not already an array) prior to adding links to the collection.