The property "Url" in "LinkPickerItem" is now resolved using either the content cache or the media cache
When the selected URL is either a content or media item, the URL of the item may have changed over time since a content item can be moved or renamed, and a media item can be replaced with a file with another filename. To make sure we reference the newest URL, the URL will now be resolved using either the content cache or the media cache. This will also make sure the correct domain is used for working with multiple sites in the same Umbraco installation.